Exploring the images
This demo shows how to explore the images generated by CRIU. As an example the image set from the "single ZDTM test run" demo.

To explore the images the CRIT command is used, which dumps the images contents in JSON format. In this demo we find what the root process of the test was, check what files it had opened and find out their names.